tailcat: skip netcheck DERP region auto-selection on js/wasm
Move PickBestRegion out of tailcat.go into its own pickregion.go and
add a js/wasm stub that always returns 0: netcheck's STUN probes need
raw UDP, which browsers don't have. Returning 0 makes ConnInfo.Expand
take its existing fallback path of picking an arbitrary region,
relying on the DERP map server to have filtered the regions by client
proximity. This also keeps netcheck out of browser wasm binaries.

This is groundwork for a browser wasm build of tailcat.

Updates tailscale/corp#24454
